% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{Imports the raster and point layers.
Creates a raster stack.
Checks the point data for inconsistency problems and removes blanks.}
\usage{
data_import(predictors, points)
}
\arguments{
\item{points}{The name of a .csv file.}

\item{rasters}{A string of raster file names.}
}
\value{
A list of objecgs predictors (a raster stack) and plots (a data.table).
}
\description{
Imports the raster and point layers.
Creates a raster stack.
Checks the point data for inconsistency problems and removes blanks.
}
\examples{
data_import(input_rasters,input_points)
}
