% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_import.R
\name{plots_import}
\alias{plots_import}
\title{Imports the raster and point layers. Creates a raster stack. Checks the point data for inconsistency problems and removes blanks.}
\usage{
plots_import(predictors, plots_data)
}
\arguments{
\item{rasters}{A string of raster file names.}

\item{points}{The name of a .csv file.}
}
\value{
A list of objecgs predictors (a raster stack) and plots (a data.table).
}
\description{
Imports the raster and point layers. Creates a raster stack. Checks the point data for inconsistency problems and removes blanks.
}
\examples{
plots_import(input_rasters,input_points)
}
