% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_import.R
\name{raster_import}
\alias{raster_import}
\title{Imports the raster and creates a raster stack}
\usage{
raster_import(rasters, extents_raster)
}
\arguments{
\item{rasters}{A string of raster file names}

\item{extents_raster}{A raster of the desired study extent.}
}
\value{
Objedt predictors - A raster stack
}
\description{
Imports the raster and creates a raster stack
}
\examples{
rasters <- c(
'data/PRISM/PRISM_ppt_30yr_normal_800mM2_annual_asc/ppt_30yr_n.asc',
'data/PRISM/PRISM_tmean_30yr_normal_800mM2_annual_asc/tmean_30yr.asc',
'data/PRISM/PRISM_us_dem_800m_asc/us_dem.asc'
)
raster_import(rasters)
}
