% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{glm_model}
\alias{glm_model}
\title{Takes the plots data.table and runs a GLM model.}
\usage{
glm_model(inputs, formula)
}
\arguments{
\item{inputs}{A string of dependent and independant variables}

\item{data}{A data.table object.}
}
\value{
A list object of a glm model object and the Psuedo R^{2} value.
}
\description{
Takes the plots data.table and runs a GLM model.
}
\examples{
glm_model(input_model, input_formula)

}
